+++ instrumentation/google-http-client-1.19/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/googlehttpclient/GoogleHttpClientAsyncTest.java
@@ -1,17 +1,23 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
 package io.opentelemetry.javaagent.instrumentation.googlehttpclient;
 
 import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpResponse;
+import io.opentelemetry.instrumentation.testing.junit.http.HttpClientTypeAdapter;
+import io.opentelemetry.instrumentation.testing.junit.http.NewHttpClientInstrumentationExtension;
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.TestFactory;
+import org.junit.jupiter.api.extension.RegisterExtension;
+import java.util.Collection;
 
-class GoogleHttpClientAsyncTest extends AbstractGoogleHttpClientTest {
+class GoogleHttpClientAsyncTest {
 
-  @Override
-  protected HttpResponse sendRequest(HttpRequest request) throws Exception {
-    return request.executeAsync().get();
+  @RegisterExtension
+  static final NewHttpClientInstrumentationExtension testing = NewHttpClientInstrumentationExtension.forAgent();
+
+  @TestFactory
+  Collection<DynamicTest> test() {
+    HttpClientTypeAdapter<HttpRequest> adapter = new GoogleClientAdapter(
+        request -> request.executeAsync().get());
+    GoogleHttpClientTests googleTests = GoogleHttpClientTests.create(adapter, testing.getTestRunner(), testing.getServer());
+    return googleTests.all();
   }
 }
