+++ instrumentation/google-http-client-1.19/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/googlehttpclient/GoogleHttpClientSyncTest.java
@@ -1,9 +1,6 @@
 package io.opentelemetry.javaagent.instrumentation.googlehttpclient;
 
 import com.google.api.client.http.HttpRequest;
-import io.opentelemetry.instrumentation.testing.junit.http.HttpClientTestOptions;
-import io.opentelemetry.instrumentation.testing.junit.http.HttpClientTestOptionsBuilder;
-import io.opentelemetry.instrumentation.testing.junit.http.HttpClientTests;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpClientTypeAdapter;
 import io.opentelemetry.instrumentation.testing.junit.http.NewHttpClientInstrumentationExtension;
 import java.util.Collection;
@@ -11,9 +8,7 @@ import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import static io.opentelemetry.javaagent.instrumentation.googlehttpclient.NewGoogleHttpClientTests.buildOptions;
-
-class NewGoogleHttpClientSyncTest {
+class GoogleHttpClientSyncTest {
 
   @RegisterExtension
   static final NewHttpClientInstrumentationExtension testing = NewHttpClientInstrumentationExtension.forAgent();
@@ -21,13 +16,8 @@ class NewGoogleHttpClientSyncTest {
   @TestFactory
   Collection<DynamicTest> test() {
     HttpClientTypeAdapter<HttpRequest> adapter = new GoogleClientAdapter(HttpRequest::execute);
-
-    HttpClientTestOptions options = buildOptions();
-    HttpClientTests<HttpRequest> clientTests = new HttpClientTests<>(testing.getTestRunner(),
-        testing.getServer(), options, adapter);
-
-    NewGoogleHttpClientTests googleTests = new NewGoogleHttpClientTests(clientTests, adapter);
-
+    GoogleHttpClientTests googleTests = GoogleHttpClientTests.create(adapter,
+        testing.getTestRunner(), testing.getServer());
     return googleTests.all();
   }
 }
