+++ instrumentation/google-http-client-1.19/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/googlehttpclient/legacy/GoogleHttpClientAsyncTest.java
@@ -3,15 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.googlehttpclient;
+package io.opentelemetry.javaagent.instrumentation.googlehttpclient.legacy;
 
 import com.google.api.client.http.HttpRequest;
 import com.google.api.client.http.HttpResponse;
 
-class GoogleHttpClientTest extends AbstractGoogleHttpClientTest {
+class GoogleHttpClientAsyncTest extends AbstractGoogleHttpClientTest {
 
   @Override
   protected HttpResponse sendRequest(HttpRequest request) throws Exception {
-    return request.execute();
+    return request.executeAsync().get();
   }
 }
