+++ instrumentation/google-http-client-1.19/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/googlehttpclient/legacy/GoogleHttpClientTest.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.googlehttpclient.legacy;
+
+import com.google.api.client.http.HttpRequest;
+import com.google.api.client.http.HttpResponse;
+
+class GoogleHttpClientTest extends AbstractGoogleHttpClientTest {
+
+  @Override
+  protected HttpResponse sendRequest(HttpRequest request) throws Exception {
+    return request.execute();
+  }
+}
